+++ src/h.js
@@ -298,9 +298,13 @@ var H = (function() {
 
     // prepare callbacks
 
+    var t0 = performance.now();
+
     r.onload = function() {
-      var o = { status: r.status, request: r };
-      o.data = null; try { o.data = JSON.parse(r.responseText); } catch (ex) {};
+      var o = {
+        status: r.status, request: r, duration: performance.now() - t0 };
+      o.data = null; try {
+        o.data = JSON.parse(r.responseText); } catch (ex) {};
       if (as.cbs.onok && r.status === 200)
         as.cbs.onok(o);
       else